refactoring/restructuring data packages
