+++ api/server/server.go
@@ -8,14 +8,14 @@ import (
 
 	// "github.com/appcelerator/amp/api/rpc/project"
 	"github.com/appcelerator/amp/api/rpc/service"
-	"github.com/appcelerator/amp/data"
-	"github.com/appcelerator/amp/data/etcd"
+	"github.com/appcelerator/amp/data/storage"
+	"github.com/appcelerator/amp/data/storage/etcd"
 	"google.golang.org/grpc"
 )
 
 var (
-	// Store is the interface used to access etcd backend
-	Store data.Store
+	// Store is the interface used to access the key/value storage backend
+	Store storage.Interface
 )
 
 // Start starts the server
