+++ data/storage/README.md
@@ -1,0 +1,5 @@
+# data.store
+
+The `data.store` packages contains the interface to key/value storage. AMP's
+primary key/value store is `etcd` and the `data.store` interface is implemented
+in the `data.store.etcd` package.
